manager: Don't switch to WorldPreview instances on SSG
The setseed manager previously treated instances on the world preview with a good spawn the same as those ready to be played with a good spawn. This allowed for the player to get switched into an instance still on the world preview screen. This commit fixes this issue; players can only get sent to instances ready to be played.