refactor: Restore basic functionality
The cycle/standard reset style is now usable once again, albeit without
affinity or reset counting (the former needs reworking, the latter is
simply not implemented as of yet.) The codebase is definitely much less
of a mess to work with now.

The departure from the manager/worker style to having one central reset
function and multiple asynchronous log readers should help to keep
things a bit less messy, with no need for the manager to continuously
"steal" the current state of each instance from the workers.