reset: Change preview pausing
Should be more consistent (if anyone was having issues with preview pausing). Should also reduce input lag on the wall, especially when running lots of instances. The pause menu can be a bit slow to pop up at times; if people have issues with pausing on entering the world as well then I think adding a timer system for both preview and idle pauses will be worth doing.