cfg/reset: Add CCX pinning for affinity
Splits the instances in half and gives half to each CCX. At least on my 5900X, it seems that the cores from each CCX are striped across one another (C0 is CCX0, C1 is CCX1, C2 is CCX0, and so on). If this turns out to be different on certain CPUs this will have to be reworked. This does give a substantial performance boost when resetting for me though.