x11: Cleanup, paper over GLFW hackfixes
I think this is just a better solution to the GLFW messiness. It removes the need for any other code to deal with timestamps in a way which is *good enough* for the current use case. This will need to be adjusted if many of the same key (e.g. tab) are ever sent continuously, such as for instance warmup or opening to LAN.