repo: handle commit descriptions properly in on-commit workflow
Adding a commit description caused `github.event.head_commit.message` to return multiple lines of content which caused curl to send invalid JSON to the Discord webhook endpoint. Escaping the newlines in the commit message to include the description is also an option but I think just using the name alone is good enough.