chore: Sync changes from `dev`. (#22)
* bench, mc: remove leave preview key support

The Leave Preview key was removed in the newest version of WorldPreview
(6.0.0, still illegal), and the easiest fix is to remove support for it
altogether. It provides little benefit on the versions where it still
exists anyway.

* mc: Fix unknown state bug.

* ctl: toggle pointer grab with wall focus hotkey

Having more ways for people to disable the pointer grab would be useful,
and being able to use the wall focus keybind to toggle it is somewhat
intuitive (people have already tried to do it, and apparently it somehow
worked for some?)

* repo: Add glfw-patch as a release artifact.

* repo: Fix libxrandr for glfw-patch artifact.

* repo: Fix upload run in on-commit action.

* cfg, ctl: add julti_locking option

Several people have expressed that they prefer the behavior of locked
instances on Julti. This new configuration option should emulate Julti's
handling of locked instances:

  - When enabled, locked instances can be reset immediately with the
    reset hotkey.
  - When enabled, pressing the lock keybind on an already locked
    instance will not unlock it.

The behavior of locked instances should be unchanged when the option is
disabled.

* ctl: unlock instance on reset

* feat: Remove wall and moving wall (RIP resetti wall).

* fix: rpm builder.

* fix: Remove OBS docs.

* fix: FindInstance styling and default config header comment.

* fix: Channel overflow bug causing x11 thread suspension.

* feat: Add focus_lost and focus_gained hooks.

---------

Co-authored-by: tesselslate <46545045+tesselslate@users.noreply.github.com>